Move ValueInto to row.rs & rename fn to just into()
