+++ src/lib.rs
@@ -294,9 +294,9 @@ mod cursor;
 mod row;
 mod statement;
 
-pub use connection::{Connection, OpenFlags, ValueInto};
+pub use connection::{Connection, OpenFlags};
 pub use cursor::Cursor;
-pub use row::{ColumnIndex, Row};
+pub use row::{ColumnIndex, Row, ValueInto};
 pub use statement::{Bindable, Readable, State, Statement};
 
 /// Open a read-write connection to a new or existing database.
